Refactor src/string - attempt om improving tree-shakeability
