+++ src/string/conditions.ts
@@ -1,0 +1,5 @@
+import { lazyStrMethod } from './guard';
+
+export const startsWith = lazyStrMethod('startsWith');
+
+export const endsWith = lazyStrMethod('endsWith');
