+++ src/string/formatting.ts
@@ -1,0 +1,3 @@
+import { replace } from './replace';
+
+export const getDigits = replace(/[^0-9]/g, '');
