+++ src/string/index.ts
@@ -1,0 +1,14 @@
+export * from './case';
+export * from './conditions';
+export * from './formatting';
+export * from './guard';
+export * from './match';
+export * from './prepend';
+export * from './replace';
+export * from './search';
+export * from './split';
+export * from './substr';
+export * from './test';
+export * from './trim';
+
+export { concat, includes } from '../generic';
