+++ src/string/prepend.ts
@@ -1,0 +1,3 @@
+import { safeStr } from './guard';
+
+export const prepend = (prefix: string) => safeStr(s => prefix.concat(s));
